Added bash script to run till convergence, other fixes in reducer and mapper

+++ runKMeans.sh
@@ -1,3 +1,14 @@
-#!/usr/bin/bash
+#!/bin/bash
 
-hdfs dfs -rm -r test/output; hadoop jar /usr/local/hadoop/share/hadoop/tools/lib/hadoop-streaming-2.8.1.jar -file mapper.py -mapper mapper.py -file reducer.py -reducer reducer.py -input ./cho.txt -output test/output
+num_itr=0
+while :
+do
+if [ ! -f 'stopIteration' ]; then
+    echo "Not converged yet. Current Iteration #: $num_itr"
+    let "num_itr++"
+    hdfs dfs -rm -r test/output; hadoop jar /usr/local/hadoop/share/hadoop/tools/lib/hadoop-streaming-2.8.1.jar -file mapper.py -mapper mapper.py -file reducer.py -reducer reducer.py -input ./cho.txt -output test/output
+    else
+        echo "Converged after $num_itr iterations! Stopping iteration"
+        break
+fi
+done
